Fix doubled zero-width spaces in data
Doubled-up ZWS characters (not NBSP as I earlier wrote by mistake)
appear not just in the dates, but also in the headlines and articles. In
fact, up to four consecutive ZWS characters appear in some parts of the
data. So we clean up any (ZWS)+ sequences to a single ZWS.